fix attach crash
If we happen to reattach right between when a server sends a buffer and when it
sends the commit message, the buffer size in the commit message will be 0 and we
should ignore it. The odds of this happening are actually not bad since buffer
and commit messages are the vast majority of the messages sent and sending a
buffer takes a few ms, a decent percentage of the time between frames.