Expand Xwayland screen size
Right now there isn't a way to handle events wayland events that
move beyond the bounds of the xwayland screen.  E.g. mouse events
outside the bounds of the wayland surface will hit the screen bounds
when passed to x11.  So let's make the xwayland screen bigger
so that we have some buffer to work with.